Finished task 2 of showing orders by year

+++ app/src/main/java/com/nefariousdream/shopifyapp/data/model/Order.java
@@ -16,7 +16,7 @@ public class Order implements Serializable
     public String totalPriceUsd;
     @SerializedName("order_number")
     @Expose
-    public Long orderNumber;
+    public int orderNumber;
     @SerializedName("contact_email")
     @Expose
     public String contactEmail;
